+++ tests/models/tests_ogc_request.py
@@ -1,9 +1,14 @@
 import os
 
+from django.db.models.expressions import Value
+from django.db.models.query_utils import Q
 from django.test import RequestFactory, SimpleTestCase
 from eulxml.xmlmap import load_xmlobject_from_file
 
 from ows_lib.models.ogc_request import OGCRequest
+from ows_lib.xml_mapper.exceptions import (
+    InvalidParameterValueException,
+    MissingConstraintLanguageParameterException)
 from ows_lib.xml_mapper.xml_requests.wfs.get_feature import GetFeatureRequest
 from tests.settings import DJANGO_TEST_ROOT_DIR
 
@@ -67,3 +72,82 @@ class OGCRequestTest(SimpleTestCase):
         self.assertEqual(["ms:Countries"], ogc_request.requested_entities)
         self.assertTrue(isinstance(ogc_request.xml_request, GetFeatureRequest))
         ogc_request.prepare()
+
+    def test_ogc_request_with_get_records_request_without_constraint_langugage(self):
+        """Test that OGCRequest helper class works correctly for a given GetMap get request"""
+
+        # Missing constraint language
+        ogc_request: OGCRequest = OGCRequest(
+            method="GET",
+            url="http://mrmap-proxy/csw/cd16cc1f-3abb-4625-bb96-fbe80dbe23e3/",
+            params={"REQUEST": ["GetRecords"], "SERVICE": "CSW", "VERSION": "2.0.2", "Constraint": "title LIKE '%ips%'"})
+
+        self.assertTrue(ogc_request.is_csw)
+        self.assertTrue(ogc_request.is_get_records_request)
+
+        self.assertEqual(
+            MissingConstraintLanguageParameterException,
+            type(ogc_request.filter_constraint()))
+
+    def test_ogc_request_with_get_records_request_with_cql_filter(self):
+        """Test that OGCRequest helper class works correctly for a given GetMap get request"""
+
+        # cql filter
+        ogc_request: OGCRequest = OGCRequest(
+            method="GET",
+            url="http://mrmap-proxy/csw/cd16cc1f-3abb-4625-bb96-fbe80dbe23e3/",
+            params={"REQUEST": ["GetRecords"], "SERVICE": "CSW", "VERSION": "2.0.2", "Constraint": "title LIKE '%ips%' ", "CONSTRAINTLANGUAGE": "CQL_TEXT"})
+
+        expected_query = Q(title__contains='ips')
+
+        self.assertTrue(ogc_request.is_csw)
+        self.assertTrue(ogc_request.is_get_records_request)
+        f = ogc_request.filter_constraint()
+        self.assertEqual(expected_query, f)
+
+    def test_ogc_request_with_get_records_request_with_wrong_cql_filter(self):
+        """Test that OGCRequest helper class works correctly for a given GetMap get request"""
+
+        # cql filter
+        ogc_request: OGCRequest = OGCRequest(
+            method="GET",
+            url="http://mrmap-proxy/csw/cd16cc1f-3abb-4625-bb96-fbe80dbe23e3/",
+            params={"REQUEST": ["GetRecords"], "SERVICE": "CSW", "VERSION": "2.0.2", "Constraint": "title like '%ips%' ", "CONSTRAINTLANGUAGE": "CQL_TEXT"})
+
+        self.assertTrue(ogc_request.is_csw)
+        self.assertTrue(ogc_request.is_get_records_request)
+        self.assertEqual(
+            InvalidParameterValueException,
+            type(ogc_request.filter_constraint()))
+
+    def test_ogc_request_with_get_records_request_with_wrong_fes_filter(self):
+        """Test that OGCRequest helper class works correctly for a given GetMap get request"""
+
+        # cql filter
+        ogc_request: OGCRequest = OGCRequest(
+            method="GET",
+            url="http://mrmap-proxy/csw/cd16cc1f-3abb-4625-bb96-fbe80dbe23e3/",
+            params={"REQUEST": ["GetRecords"], "SERVICE": "CSW", "VERSION": "2.0.2", "Constraint": '<ogc:Filter xmlns:ogc="http://www.opengis.net/ogc"><ogc:a><ogc:ValueReference>type</ogc:ValueReference><ogc:Literal>dataset</ogc:Literal></ogc:a></ogc:Filter>', "CONSTRAINTLANGUAGE": "FILTER"})
+
+        self.assertTrue(ogc_request.is_csw)
+        self.assertTrue(ogc_request.is_get_records_request)
+
+        self.assertEqual(
+            InvalidParameterValueException,
+            type(ogc_request.filter_constraint()))
+
+    def test_ogc_request_with_get_records_request_with_fes_filter(self):
+        """Test that OGCRequest helper class works correctly for a given GetMap get request"""
+
+        # cql filter
+        ogc_request: OGCRequest = OGCRequest(
+            method="GET",
+            url="http://mrmap-proxy/csw/cd16cc1f-3abb-4625-bb96-fbe80dbe23e3/",
+            params={"REQUEST": ["GetRecords"], "SERVICE": "CSW", "VERSION": "2.0.2", "Constraint": '<ogc:Filter xmlns:ogc="http://www.opengis.net/ogc"><ogc:PropertyIsEqualTo><ogc:ValueReference>type</ogc:ValueReference><ogc:Literal>dataset</ogc:Literal></ogc:PropertyIsEqualTo></ogc:Filter>', "CONSTRAINTLANGUAGE": "FILTER"})
+
+        expected_query = Q(type__exact=Value('dataset'))
+
+        self.assertTrue(ogc_request.is_csw)
+        self.assertTrue(ogc_request.is_get_records_request)
+        f = ogc_request.filter_constraint()
+        self.assertEqual(expected_query, f)